drivers: ptp_clock: use signed ppb rate adjustment
Change the PTP clock driver rate_adjust callback from a floating-point nominal-frequency ratio to a signed parts-per-billion correction.

Update the in-tree PTP and gPTP callers, driver implementations, tests, and migration guide to use the new ppb semantics.

+++ subsys/net/l2/ethernet/gptp/gptp_private.h
@@ -125,9 +125,9 @@ static inline uint64_t gptp_timestamp_to_nsec(struct net_ptp_time *ts)
  *
  * @param nanosecond_diff nanosecond offset.
  *
- * @return ppb value to adjust.
+ * @return signed ppb value to adjust.
  */
-double gptp_servo_pi(int64_t nanosecond_diff);
+int32_t gptp_servo_pi(int64_t nanosecond_diff);
 
 /**
  * @brief Change the port state
